move logging and types to util to make python's module resolution happierq
