+++ src/levanter/__init__.py
@@ -3,7 +3,6 @@ import levanter.config as config
 import levanter.data as data
 import levanter.distributed as distributed
 import levanter.eval as eval
-import levanter.logging as logging
 import levanter.models as models
 import levanter.optim as optim
 import levanter.tracker as tracker
