+++ src/levanter/callbacks.py
@@ -19,12 +19,12 @@ from tqdm_loggable.auto import tqdm
 
 import levanter.tracker
 from levanter.data import AsyncDataset, DataLoader
-from levanter.logging import save_xla_dumps_to_wandb
 from levanter.tracker.helpers import log_optimizer_hyperparams
 from levanter.tracker.wandb import WandbConfig
 from levanter.trainer import StepInfo
 from levanter.utils import flop_utils
 from levanter.utils.jax_utils import barrier_sync, jnp_to_python
+from levanter.utils.logging import save_xla_dumps_to_wandb
 from levanter.visualization import compute_and_visualize_log_probs as viz_probs
