+++ src/levanter/checkpoint.py
@@ -26,7 +26,7 @@ import haliax.partitioning
 from haliax.jax_utils import is_in_jit, is_jax_array_like
 
 from levanter.tensorstore_serialization import tree_deserialize_leaves_tensorstore, tree_serialize_leaves_tensorstore
-from levanter.types import FilterSpec
+from levanter.utils.types import FilterSpec
 
 
 logger = logging.getLogger(__name__)
