+++ src/levanter/eval.py
@@ -17,10 +17,10 @@ from haliax.partitioning import ResourceMapping
 
 import levanter.tracker
 from levanter.data import AsyncDataset, DataLoader
-from levanter.logging import LoadingTimeTrackerIterator
 from levanter.models.lm_model import LmExample, LmHeadModel, compute_next_token_loss
 from levanter.trainer import StepInfo
 from levanter.utils.hf_utils import HfTokenizer, byte_length_of_token
+from levanter.utils.logging import LoadingTimeTrackerIterator
 from levanter.utils.stat_utils import Arrayish, RunningMean
 from levanter.utils.tree_utils import inference_mode
