+++ src/levanter/main/cache_dataset.py
@@ -6,9 +6,9 @@ import levanter
 from levanter.data.metrics_monitor import LoggingMetricsMonitor, RichMetricsMonitor
 from levanter.data.text import BatchTokenizer, LMDatasetConfig
 from levanter.distributed import RayConfig
-from levanter.logging import init_logging
 from levanter.store.cache import build_or_load_cache
 from levanter.tracker import NoopConfig, TrackerConfig
+from levanter.utils.logging import init_logging
 
 
 logger = logging.getLogger(__name__)
